requirements.txt
python-dotenv
requests
mypy
types-requests
pydantic
